requirements.txt
boto3
websocket-client